requirements.txt
Flask==3.0.0
pdfminer.six==20240706
Werkzeug>=3.0.0
